Changes to reduce the negative impact of csshover.htc on IE performance when page contains large a javascript generated DOM component (ie. popups in gbrowse imagemap). Largest impact appears to be flexigrid.css triggering csshover.htc to crawl through the imagemap and attach events to non-anchor elements (tr:hover). So, moving flexigrid.css link out of the header.tag and into just customSummary.jsp (under the belief that this is the only page that uses it).
Also changing '.detail_link:hover' to 'a.detail_link:hover' as that added a small performance improvement for IE on growse-containg pages.

There is still a processing delay due to 'li:hover's in the ${project}.css but those are required for header menus. IE7 shouldn't need csshover.htc as it has native support for non-anchor hovers. The menubar menus are fine in IE7 w/o the htc but the About Us, etc menus currently require the htc. 



git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/apidb/ApiCommonWebsite/trunk@27470 735e2a04-f8fc-0310-8a1b-f2942603c481